2005-05-07  H.J. Lu  <hongjiu.lu@intel.com>
	* d10v-dis.c (dis_2_short): Support 64bit host.

+++ opcodes/d10v-dis.c
@@ -253,7 +253,8 @@ dis_2_short (insn, memaddr, info, order)
       while (op->name)
 	{
 	  if ((op->format & SHORT_OPCODE)
-	      && ((op->mask & ins[j]) == (unsigned long) op->opcode))
+	      && ((((unsigned int) op->mask) & ins[j])
+		  == (unsigned int) op->opcode))
 	    {
 	      (*info->fprintf_func) (info->stream, "%s\t", op->name);
 	      for (i = 0; op->operands[i]; i++)
